Seminar_app/requirements.txt
django
pillow
numpy
pandas
opencv-python-headless
torch
torchvision